requirements.txt
numpy
rich
faiss-cpu
h5py
